requirements.txt
# Core Flask
Flask==3.1.1
Flask-CORS==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1

# Database
SQLAlchemy==2.0.41
psycopg2-binary==2.9.9

# Payment Processing
stripe==11.2.0

# SMS & Communications
twilio==9.3.7

# Email
sendgrid==6.11.0

# Environment & Config
python-dotenv==1.0.1

# Validation
marshmallow==3.23.2
email-validator==2.2.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Production Server
gunicorn==23.0.0

# Testing
pytest==8.3.4
pytest-flask==1.3.0

# Existing dependencies
blinker==1.9.0
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
typing_extensions==4.14.0
Werkzeug==3.1.3
